requirements.txt
pytesseract==0.3.13
streamlit==1.41.1
PyMuPDF==1.25.1
plotly==5.24.1
Pillow==11.0.0 